Add AnalysisRunner::run function
This gives us a location in the library to run an analysis from,
rather than trying to recreate the logic that was in main.cpp if we
want to run an analysis.
