+++ src/analysisrunner.h
@@ -1,0 +1,10 @@
+#ifndef ANALYSISRUNNER_H
+#define ANALYSISRUNNER_H
+
+struct Options;
+
+namespace AnalysisRunner {
+bool run(const Options &options);
+}
+
+#endif // ANALYSISRUNNER_H
